feat(component): reorganize folder, add events
