+++ web-components/src/[sandbox]/examples/draggable-item.ts
@@ -1,7 +1,9 @@
-import "@/components/card/Card";
+import "@/components/draggable/DraggableItem";
 import { html } from "lit-element";
 
 export const draggableTemplate = html`
   <h3>Default</h3>
-  <md-draggable-item>  </md-draggable-item>
+  <md-draggable-item>A</md-draggable-item>
+  <md-draggable-item>b</md-draggable-item>
+  <md-draggable-item>C</md-draggable-item>
 `;
