+++ web-components/src/components/draggable/scss/draggable-item.scss
