+++ web-components/src/index.ts
@@ -29,7 +29,7 @@ export { DatePickerDay } from "./components/datepicker/datepicker-day/DatePicker
 export { DatePickerMonth } from "./components/datepicker/datepicker-month/DatePickerMonth";
 export { DatePickerWeek } from "./components/datepicker/datepicker-week/DatePickerWeek";
 export { DateTimePicker } from "./components/date-time-picker/DateTimePicker";
-export { DraggableItem } from "./components/draggable-item/DraggableItem";
+export { DraggableItem } from "./components/draggable/DraggableItem";
 export { EditableTextfield } from "./components/editable-textfield/EditableTextfield";
 export { FloatingModal } from "./components/floating-modal/FloatingModal";
 export { HelpText } from "./components/help-text/HelpText";
